Add files via upload
This is a video game created as an assignment from my intermediate level C++ class at City College of San Francisco with Professor Dave Harden. The program is an example of inheritance and details how a base class can have private data members and member functions which can be inherited by each of the derived classes. In the end the class objects go through a 'battle arena,' where objects simulate a fantasy battle between Humans, Elves, Cyberdemons, and Balrogs.