Add files via upload
These are the files for Project 1 from my STA 108 class with Prof. Chen Fall '17 at University of California, Davis.

A set of data which had an x and y variable which corresponded with country GDP per person plotted against the fertility rate for that country was given. The goal was to find the best way to transform the data such that it could obtain an ideal R² value. Despite the ideal transformation appearing to be the log of both the x and y plotted against each other, it was later understood that the best possible R² could be found using only 1/log(x) against the untransformed y data.

After transforming the data, it was possible to reattempt the regression analysis on the changed version of the data, allowing for confidence intervals to be developed and retested back on the original data.