Add files via upload
The final project for my STA 108 Linear Regression course with Prof. Chen involved utilizing a set of data related to diabetes patients and identifying the different variables which best predicted that a person in the population would have diabetes. Doing this meant performing a transformation on an indicator variable with BoxCox and then regressing it on all the other variables. The data was then split into training and validation data for cross-validation. Techniques that were used included "best" subset, AIC, BIC, Cp, R squared, Adjusted R squared, SSE, and the forward stepwise procedure. Multiple models were created to test different prediction criterion, and in the end, a model was selected according to it having the lowest MSPR value. Most of the work was performed in RStudio, and a report of the data along with commentary was done using Microsoft Word.